feat: 过滤函数等无法配置的属性&类型 #126955259 # Reviewed, transaction id: 62945
