+++ site/lib/client/src/components/render-edit-component/config/name-tip/index.postcss
@@ -3,8 +3,9 @@
 }
 .config-item-name-tip {
   display: inline-block;
-  padding: 1px 0;
+  margin-bottom: 2px;
   cursor: pointer;
+  font-weight: bold;
   &:hover {
     border-radius: 2px;
     color: #313238;
