+++ site/lib/client/src/components/render-edit-component/config/slot/index.postcss
@@ -16,11 +16,25 @@
   width: 400px;
   padding: 12px;
   .slot-tip-bg {
+    display: flex;
+    align-items: center;
     line-height: 24px;
+    padding: 0 12px;
+    & > span:first-child {
+      width: 48px;
+      color: #979BA5;
+      margin-right: 3px;
+    }
+    & > span:last-child {
+      flex: 1;
+      color: #4D4F56;
+      overflow: hidden;
+      text-overflow: ellipsis;
+      white-space: nowrap;
+    }
+  }
+  .even-bg {
     background-color: #F5F7FA;
     border-radius: 2px;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
   }
 }
