+++ site/lib/client/src/components/render-edit-component/dynamic-config-item/index.postcss
@@ -1,11 +1,13 @@
 .config-item {
-    margin-bottom: 16px;
+    margin-bottom: 20px;
+    &.name-value-flex {
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+    }
     &-name {
         color: #313238;
         line-height: 20px;
-    }
-    &-name.mb4 {
-        margin-bottom: 4px;
     }
     &-content {
         font-size: 0;
