+++ site/lib/client/src/components/render-edit-component/dynamic-config-item/utils.ts
@@ -1,4 +1,4 @@
-import type { IComponentWiki } from '@/types/component';
+import type { IComponentWiki, PropItem } from '@/types/component';
 
 export const basicTypeToDefVal = {
   'string': '',
@@ -9,7 +9,7 @@ export const basicTypeToDefVal = {
 }
 
 export const splitType = (type: string) => {
-  return type.split('|').map(item => item.trim().replace(/'/g, '')).filter(item => item)
+  return type.split(/\s*\|\s*/).map(item => item.trim().replace(/'/g, '')).filter(item => item)
 }
 
 export const isBasicTypeArr = (type: string) => {
@@ -70,3 +70,26 @@ export const isNumber = (value: unknown) => typeof value === 'number' || value i
 export const isBoolean = (value: unknown) => typeof value === 'boolean' || value instanceof Boolean;
 export const isArray = (value: unknown) => Array.isArray(value)
 export const isObject = (value: unknown) => Object.prototype.toString.call(value) === '[object Object]';
+
+// 暂未支持的可配置过滤掉
+export const filterErrTypeProps = (props: PropItem[], types: IComponentWiki['types']) => {
+  const partValidTypeProps = (props ?? []).filter((item: PropItem) => {
+    const typeArr = [...new Set(splitType(item.type))];
+    const factTypeList = typeArr.map((typeVal) => {
+      return factType(typeVal, item.options, types);
+    });
+    return !factTypeList.every(factType => factType === 'errortype');
+  });
+  return partValidTypeProps.map((item: PropItem) => {
+    const typeArr = [...new Set(splitType(item.type))];
+    if (typeArr.length === 1) {
+      return item;
+    }
+    const validTypes = typeArr.filter((typeValF) => {
+      const curFactType = factType(typeValF, item.options, types);
+      return curFactType !== 'errortype';
+    });
+    item.type = validTypes.join(' |');
+    return item;
+  });
+};
